mcp-jira-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-jira-server"
version = "1.0.0"
description = "Model Context Protocol (MCP) server for Jira integration"
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your-email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=23.0", "flake8>=6.0", "mypy>=1.0"]

[project.scripts]
mcp-jira-server = "mcp_jira_server.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_jira_server*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true